Utils * move prepare_files to utils * add docstryng comments
