+++ lookout/style/format/rule_stat.py
@@ -9,7 +9,7 @@ from lookout.style.format.utils import profile
 from lookout.style.format.feature_extractor import FeatureExtractor
 from lookout.style.format.feature_utils import CLASSES
 from lookout.style.format.model import FormatModel
-from lookout.style.format.quality_report import prepare_files
+from lookout.style.format.utils import prepare_files
 
 
 class RuleStat:
